refactor(hooks): 도메인별 훅 구조 정리 및 중복 제거
Related to #56
