+++ src/pages/(auth)/login/hooks/mutations/useLoginMutation.ts
@@ -1,10 +1,10 @@
 import { useMutation } from '@tanstack/react-query';
 import { useNavigate } from 'react-router-dom';
-import type { LoginPayload } from '../types';
+import type { LoginPayload } from '../../types';
 import { ROUTES } from '@/constants';
 import { useAuthStore } from '@/stores/auth';
 import { useLoadUserProfile } from '@/hooks';
-import { loginService } from '../services';
+import { loginService } from '../../services';
 
 /**
  * 로그인 뮤테이션 훅
@@ -21,8 +21,9 @@ export const useLoginMutation = () => {
     mutationFn: (payload: LoginPayload) => loginService(payload),
     onSuccess: async (result) => {
       // 액세스 토큰을 메모리에 저장 (리프레시 토큰은 HttpOnly 쿠키로 자동 관리)
-      if (result?.accessToken) {
-        setAccessToken(result.accessToken);
+      if ((result as { accessToken?: string })?.accessToken) {
+        const accessToken = (result as { accessToken: string }).accessToken;
+        setAccessToken(accessToken);
 
         try {
           // 로그인 직후 프로필 쿼리 refetch로 동기화
