+++ src/pages/(auth)/signup/hooks/mutations/useSignupMutation.ts
@@ -2,7 +2,7 @@ import { useMutation } from '@tanstack/react-query';
 import { signupService } from '../../services';
 import type { SignupPayload } from '../../types';
 
-export const useSignupMutationApi = () => {
+export const useSignupMutation = () => {
   return useMutation({
     mutationFn: async (payload: SignupPayload) => signupService.signup(payload),
   });
